Modifies embeddedFormCollections to allow filtering of fields (i.e., columns in a table) or data (i.e., rows in a table) by expressions or functions defined in the FDN. This is supported in the FDN by the following syntax:
```
collections : {
  config : {
    fieldFilter : {
      listener  : "watchListenerExpression()",
      expression: "watchExpressionOnChange(fields)"
    },
    dataFilter : {
      listener  : "watchListenerExpression()",
      expression: "watchExpressionOnChange(data)"
    }
  }
}
```

The `listener` is watched by the "embeddedFormCollection" scope, and `expression` is invoked when the `listener` has changed. The fieldFilter is invoked with the full collection's fields list and should return a filtered array from this list. The dataFilter is invoked with the collection's full data row model and should return a filtered array of this row set.